requirements.txt
streamlit==1.41.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
numpy
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
torch==2.5.1
torchvision==0.20.1
transformers==4.48.1
tf_keras==2.18.0
timm==1.0.14
requests==2.32.3
pillow==11.1.0
mediapipe==0.10.20
keras==3.8.0